Moved make_env and make_dict_env to sheeprl.utils.env
