+++ howto/register_new_algorithm.md
@@ -75,7 +75,7 @@ from sheeprl.models.models import MLP
 from sheeprl.utils.metric import MetricAggregator
 from sheeprl.utils.parser import HfArgumentParser
 from sheeprl.utils.registry import register_algorithm
-from sheeprl.utils.utils import make_env
+from sheeprl.utils.env import make_env
 
 
 def train(
